fix: migrate to Int primitive type
Notification's progress bar accepts Int progress values only.
Since computation manager sends Float progress values between 0 and 1
to be directly consumed by the UI and increase precision, convert
it to a percentage before passing it to the notification.

+++ core/testing/src/main/kotlin/com/archsoftware/afoil/core/testing/notifications/TestSystemTrayNotifier.kt
@@ -4,7 +4,7 @@ import android.app.Notification
 import com.archsoftware.afoil.core.notifications.Notifier
 
 class TestSystemTrayNotifier : Notifier {
-    override fun updateComputationServiceNotification(progress: Float) {}
+    override fun updateComputationServiceNotification(progress: Int) {}
 
     override fun createComputationServiceNotification(computationName: String?): Notification {
         return Notification()
